Updated some files

+++ src/librapid/array/helpers/extent.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
 namespace librapid {
-	template<typename T, int64_t maxDims, int64_t align_ = 1>
+	template<typename T, i32 maxDims, i32 align_ = 1>
 	class ExtentType {
 	public:
 		using Type = T;
-		static constexpr int64_t MaxDims = maxDims;
-		static constexpr int64_t Align = align_;
+		static constexpr i32 MaxDims = maxDims;
+		static constexpr i32 Align = align_;
 
 		ExtentType() = default;
 
@@ -18,7 +18,7 @@ namespace librapid {
 			LR_ASSERT(args.size() <= maxDims,
 					  "A maximum of {} dimensions are allowed in an Extent object",
 					  maxDims);
-			int64_t i = 0;
+			i32 i = 0;
 			for (const auto &val : args) m_data[i++] = val;
 		}
 
@@ -27,10 +27,10 @@ namespace librapid {
 			LR_ASSERT(args.size() <= maxDims,
 					  "A maximum of {} dimensions are allowed in an Extent object",
 					  maxDims);
-			for (int64_t i = 0; i < m_dims; ++i) m_data[i] = args[i];
+			for (i32 i = 0; i < m_dims; ++i) m_data[i] = args[i];
 		}
 
-		template<typename T_, int64_t d_, int64_t a_>
+		template<typename T_, i32 d_, i32 a_>
 		ExtentType(const ExtentType<T_, d_, a_> &e) {
 			LR_ASSERT(e.dims() < maxDims,
 					  "Extent with {} dimensions cannot be stored in an extent with a maximum of "
@@ -38,17 +38,17 @@ namespace librapid {
 					  d_,
 					  maxDims);
 			m_dims = e.dims();
-			for (int64_t i = 0; i < m_dims; ++i) { m_data[i] = e[i]; }
+			for (i32 i = 0; i < m_dims; ++i) { m_data[i] = e[i]; }
 		}
 
 		ExtentType &operator=(const ExtentType &other) {
 			if (this == &other) return *this;
 			m_dims = other.dims();
-			for (int64_t i = 0; i < m_dims; ++i) { m_data[i] = other[i]; }
+			for (i32 i = 0; i < m_dims; ++i) { m_data[i] = other[i]; }
 			return *this;
 		}
 
-		template<typename T_, int64_t d_, int64_t a_>
+		template<typename T_, i32 d_, i32 a_>
 		ExtentType &operator=(const ExtentType<T_, d_, a_> &other) {
 			LR_ASSERT(other.dims() < maxDims,
 					  "Extent with {} dimensions cannot be stored in an extent with a maximum of "
@@ -56,11 +56,11 @@ namespace librapid {
 					  d_,
 					  maxDims);
 			m_dims = other.dims();
-			for (int64_t i = 0; i < m_dims; ++i) { m_data[i] = other[i]; }
+			for (i32 i = 0; i < m_dims; ++i) { m_data[i] = other[i]; }
 			return *this;
 		}
 
-		static ExtentType zero(int64_t dims) {
+		static ExtentType zero(i32 dims) {
 			// Data is already zeroed
 			ExtentType res;
 			res.m_dims = dims;
@@ -69,8 +69,8 @@ namespace librapid {
 
 		ExtentType stride() const {
 			ExtentType res = zero(m_dims);
-			int64_t prod   = 1;
-			for (int64_t i = m_dims - 1; i >= 0; --i) {
+			i32 prod   = 1;
+			for (i32 i = m_dims - 1; i >= 0; --i) {
 				res[i] = prod;
 				prod *= m_data[i];
 			}
@@ -79,8 +79,8 @@ namespace librapid {
 
 		ExtentType strideAdjusted() const {
 			ExtentType res = zero(m_dims);
-			int64_t prod   = 1;
-			for (int64_t i = m_dims - 1; i >= 0; --i) {
+			i32 prod   = 1;
+			for (i32 i = m_dims - 1; i >= 0; --i) {
 				res[i] = prod;
 				prod *= adjusted(i);
 			}
@@ -101,7 +101,7 @@ namespace librapid {
 
 			T res			   = 0;
 			ExtentType strides = stride();
-			for (int64_t i = 0; i < index.dims(); ++i) {
+			for (i32 i = 0; i < index.dims(); ++i) {
 				res += strides[i] * index[i];
 			}
 			return res;
@@ -121,7 +121,7 @@ namespace librapid {
 
 			T res			   = 0;
 			ExtentType strides = strideAdjusted();
-			for (int64_t i = 0; i < index.dims(); ++i) {
+			for (i32 i = 0; i < index.dims(); ++i) {
 				LR_ASSERT(index.m_data[i] >= 0 && index[i] <= adjusted(i),
 						  "Index {} is out of range for Extent with adjusted dimension {}",
 						  index[i],
@@ -131,25 +131,25 @@ namespace librapid {
 			return res;
 		}
 
-		ExtentType reverseIndex(int64_t index) const {
+		ExtentType reverseIndex(i32 index) const {
 			ExtentType res	   = zero(m_dims);
 			ExtentType strides = stride();
-			for (int64_t i = 0; i < m_dims; ++i) {
+			for (i32 i = 0; i < m_dims; ++i) {
 				res[i] = index / strides[i];
 				index -= strides[i] * res[i];
 			}
 			return res;
 		}
 
-		ExtentType partial(int64_t start = 0, int64_t end = -1) const {
+		ExtentType partial(i32 start = 0, i32 end = -1) const {
 			if (end == -1) end = m_dims - 1;
 			ExtentType res;
 			res.m_dims = m_dims - 1;
-			for (int64_t i = start; i < end + 1; ++i) { res[i - start] = m_data[i]; }
+			for (i32 i = start; i < end + 1; ++i) { res[i - start] = m_data[i]; }
 			return res;
 		}
 
-		template<typename T_ = T, int64_t d = maxDims, int64_t a = align_>
+		template<typename T_ = T, i32 d = maxDims, i32 a = align_>
 		LR_NODISCARD("")
 		ExtentType swivelled(const ExtentType<T_, d, a> &order) const {
 			LR_ASSERT(
@@ -158,9 +158,9 @@ namespace librapid {
 
 #if defined(LIBRAPID_DEBUG)
 			// Check the order contains only valid numbers
-			for (int64_t i = 0; i < order.dims(); ++i) {
+			for (i32 i = 0; i < order.dims(); ++i) {
 				bool found = false;
-				for (int64_t j = 0; j < order.dims(); ++j) {
+				for (i32 j = 0; j < order.dims(); ++j) {
 					if (order[j] == i) {
 						found = true;
 						break;
@@ -171,30 +171,30 @@ namespace librapid {
 #endif
 
 			ExtentType res = zero(m_dims);
-			for (int64_t i = 0; i < order.dims(); ++i) { res[order[i]] = m_data[i]; }
+			for (i32 i = 0; i < order.dims(); ++i) { res[order[i]] = m_data[i]; }
 			return res;
 		}
-#
-		template<typename T_ = T, int64_t d = maxDims, int64_t a = align_>
+
+		template<typename T_ = T, i32 d = maxDims, i32 a = align_>
 		void swivel(const ExtentType<T_, d, a> &order) {
 			*this = swivelled(order);
 		}
 
-		LR_NODISCARD("") LR_FORCE_INLINE int64_t size() const {
-			int64_t res = 1;
-			for (int64_t i = 0; i < m_dims; ++i) res *= m_data[i];
+		LR_NODISCARD("") LR_FORCE_INLINE i32 size() const {
+			i32 res = 1;
+			for (i32 i = 0; i < m_dims; ++i) res *= m_data[i];
 			return res;
 		}
 
-		LR_NODISCARD("") LR_FORCE_INLINE int64_t sizeAdjusted() const {
-			int64_t res = 1;
-			for (int64_t i = 0; i < m_dims; ++i) res *= adjusted(i);
+		LR_NODISCARD("") LR_FORCE_INLINE i32 sizeAdjusted() const {
+			i32 res = 1;
+			for (i32 i = 0; i < m_dims; ++i) res *= adjusted(i);
 			return res;
 		}
 
-		LR_NODISCARD("") LR_FORCE_INLINE int64_t dims() const { return m_dims; }
+		LR_NODISCARD("") LR_FORCE_INLINE i32 dims() const { return m_dims; }
 
-		const T &operator[](int64_t index) const {
+		const T &operator[](i32 index) const {
 			LR_ASSERT(index >= 0 && index < m_dims,
 					  "Index {} is out of range for Extent with {} dimensions",
 					  index,
@@ -202,7 +202,7 @@ namespace librapid {
 			return m_data[index];
 		}
 
-		T &operator[](int64_t index) {
+		T &operator[](i32 index) {
 			LR_ASSERT(index >= 0 && index < m_dims,
 					  "Index {} is out of range for Extent with {} dimensions",
 					  index,
@@ -210,7 +210,7 @@ namespace librapid {
 			return m_data[index];
 		}
 
-		T adjusted(int64_t index) const {
+		T adjusted(i32 index) const {
 			LR_ASSERT(index >= 0 && index < m_dims,
 					  "Index {} is out of range for Extent with {} dimensions",
 					  index,
@@ -218,18 +218,18 @@ namespace librapid {
 			return roundUpTo(m_data[index], Align);
 		}
 
-		template<typename T_, int64_t d_, int64_t a_>
+		template<typename T_, i32 d_, i32 a_>
 		LR_NODISCARD("")
 		bool operator==(const ExtentType<T_, d_, a_> &other) const {
 			if (m_dims != other.m_dims) return false;
-			for (int64_t i = 0; i < m_dims; ++i)
+			for (i32 i = 0; i < m_dims; ++i)
 				if (m_data[i] != other.m_data[i]) return false;
 			return true;
 		}
 
 		LR_NODISCARD("") std::string str() const {
 			std::string res = "Extent(";
-			for (int64_t i = 0; i < m_dims - 1; ++i) res += fmt::format("{}, ", m_data[i]);
+			for (i32 i = 0; i < m_dims - 1; ++i) res += fmt::format("{}, ", m_data[i]);
 			return res + fmt::format("{})", m_data[m_dims - 1]);
 		}
 
@@ -242,8 +242,8 @@ namespace librapid {
 					  index,
 					  str());
 
-			int64_t extentProd = 1;
-			for (int64_t i = index + 1; i < m_dims; ++i) extentProd *= adjusted(i);
+			i32 extentProd = 1;
+			for (i32 i = index + 1; i < m_dims; ++i) extentProd *= adjusted(i);
 			return extentProd * first;
 		}
 
@@ -255,8 +255,8 @@ namespace librapid {
 					  index,
 					  str());
 
-			int64_t extentProd = 1;
-			for (int64_t i = index + 1; i < m_dims; ++i) extentProd *= adjusted(i);
+			i32 extentProd = 1;
+			for (i32 i = index + 1; i < m_dims; ++i) extentProd *= adjusted(i);
 			return extentProd * first + indexImpl(index + 1, others...);
 		}
 
@@ -268,8 +268,8 @@ namespace librapid {
 					  index,
 					  str());
 
-			int64_t extentProd = 1;
-			for (int64_t i = index + 1; i < m_dims; ++i) extentProd *= adjusted(i);
+			i32 extentProd = 1;
+			for (i32 i = index + 1; i < m_dims; ++i) extentProd *= adjusted(i);
 			return extentProd * first;
 		}
 
@@ -281,8 +281,8 @@ namespace librapid {
 					  index,
 					  str());
 
-			int64_t extentProd = 1;
-			for (int64_t i = index + 1; i < m_dims; ++i) extentProd *= adjusted(i);
+			i32 extentProd = 1;
+			for (i32 i = index + 1; i < m_dims; ++i) extentProd *= adjusted(i);
 			return extentProd * first + indexImpl(index + 1, others...);
 		}
 
@@ -291,11 +291,11 @@ namespace librapid {
 		T m_data[maxDims] {};
 	};
 
-	template<typename T, int64_t d, int64_t a>
+	template<typename T, i32 d, i32 a>
 	inline std::string str(const ExtentType<T, d, a> &val,
 						   const StrOpt &options = DEFAULT_STR_OPT) {
 		return val.str();
 	}
 
-	using Extent = ExtentType<int64_t, 32, 1>;
+	using Extent = ExtentType<i32, 32, 1>;
 } // namespace librapid
